+++ crates/mimic_declare/src/node/entity.rs
@@ -119,15 +119,32 @@ impl AsType for Entity {
     }
 
     fn as_view_type(&self) -> TokenStream {
+        let derives = Self::basic_derives();
         let view_ident = &self.def.view_ident();
         let view_field_list = AsType::as_view_type(&self.fields);
-        let derives = Self::basic_derives();
+        let view_default = self.view_default();
 
         // quote
         quote! {
             #derives
             pub struct #view_ident {
                 #view_field_list
+            }
+            #view_default
+        }
+    }
+
+    fn view_default(&self) -> TokenStream {
+        let view_ident = &self.def.view_ident();
+        let view_defaults = self.fields.view_default();
+
+        quote! {
+            impl Default for #view_ident {
+                fn default() -> Self {
+                    Self {
+                        #view_defaults
+                    }
+                }
             }
         }
     }