+++ crates/mimic_declare/src/traits.rs
@@ -109,6 +109,8 @@ pub trait AsType {
 
     fn as_view_type(&self) -> TokenStream;
 
+    fn view_default(&self) -> TokenStream;
+
     fn basic_derives() -> TokenStream {
         TraitList(vec![
             Trait::CandidType,